Done with the first example.
As expected, the parser was by far the hardest part of this whole program. It took me a while.

I've managed to reuse the GUI bindings from my poker project suprisingly enough and rather than do that command line parsing that the Ocaml examples, have I've decided to slap a GUI on top.

This thing is slow going so far, but I will be able to reuse what I did here in the following examples.

I won't bother translating pretty printing, command line parsing and as can be seen here I've decided to replace Lex and Yacc's generators with Fparsec's parser combinators.