Change of plans, rather than translate all the examples, I'll make this an experimental repo for the examples I am interested in.
I did not expect it to be so much code in `fulluntyped`, nor the fact that in the test cases I actually noticed assignment, so that wrecks the model I had in the Arith example. I do not feel like spending weeks just writing parsers which is what would be the case should I go with the plan to do every example.

But going through the notions did make what the code does a lot clearer to me than just reading the book.

Let me move to the next thing.