No, I can't go on anymore.
I will take a break here and move on to the Haskell paper after lunch. I do not really like Pierce's code, but as a last ditch effort, I will translate it to F# if necessary.

The small step semantics of constraint typing are pretty confusing here unless you already know what the algorithm is supposed to be about which I don't. I'll take the advice offered to me on the types subreddit and move on to bidirectional search.