I've gone through the Haskell examples by Diehl and Grubmiller and my conclussion is that the use of typeclasses and monads make the code significantly more clearer that the SML version from the first tutorial.
So if I am going to learn this thing, it would be a good idea to do it in Haskell.

That is what I am going to go with.

The point of this whole exercise in the end is to find hope - it might be the case that the DU style is the most flexible and powerful one given the level of type inference available to current functional languages.

Therefore, the way to recover lost safety is to do type inference programatically within the confines of the DSL rather than the main language.

I strongly suspect that either way, I am going to be structuring my most advanced programs around pattern matching. If so, I either resolve myself to writting a dynamic language that does type inference the way dynamic languages do...or I do what I am doing now to completion and internalize type inference algorithms.