Equality constraints should be (Expr, Type) right? I won't peek.
Just like I did it for the Spiral library, reading books or reading code will only get me so far. I feel like struggling with this for a while to train my memory.

Yeah, now that I did it like this, I finally understand how little I really understand.

It is like I did nothing for the past two weeks. In order to internalize this, I need to face down the wall in front of me. That having said, as I am getting tired and it is getting late, I'll try a few more things and call it day.