A few things come to might:
1) Algorithm W has such an intricate and complex structure, I can't make any headway on internalizing it at all.

2) I am having strong doubts that Algorithm W is the way to go because of the above. Can something this complex really be the way to do type inference? My brief experience with search methods tells me that there are numerous ways to cut the pie; this can't really be it.

3) I am pretty much stuck right now. What should I do next? Well I suppose I could translate the SML example word for word to F#, but I am not sure there is a point to this.

This situation reminds of the one I was back in high school.

I coasted along find in middle school, but the difficult programming contest problems were way too much for me in high school. Even with the solutions provided with code, I could not make heads or tails of the problems given to me.

This pretty much crushed me as programmer and decided to go into trading instead which did not pan out either. Well, that last one did not end too badly as I discovered the meaning of life in the Singularity, but

I'd say it is important to believe, but there is nothing emptier in life than belief that is not reciprocated. In the end, small steps need to provide some benefit or they won't turn into larger ones.

Maybe I'll do another change of plans and do some research into bidirectional search. Let me go with that.