I finished translating Oleg's example to F# and testing it to boot. Before finishing the job, I want to convert the example to a monadic Haskell one with the state and the exception monads.
The way exceptions are used for errors in the F# example is kind of a pain in the ass.

I wrote that monads were a pain yesterday, but I doubt they would be worse than converting this whole thing to use its own combinators.

Let me do it and then I will take a look at the other Algorithm W examples.