I am 2/3rds into Diehl tutorial. To be honest, I will call it a day here. I do not have the energy to finish this.
Shuffling those substitutions around is a really bad idea, Oleg's design seems a lot more sound to me.

Given how this turned out, I am guessing that separating constraint generation from the rest of program might be a better of going about it.

That having said, the main difference between Oleg's and Diehl's examples are that Oleg did not bother with let-polymorphism. If he did, it might have been the case that his example would be just as complex as Diehl's.

Regardless, I still am suspicious about that substitution passing. I'll figure it out tomorrow and then go through that final, third example. After that it is time to go into bidirectional checking.

Going through this line by line has given me the confidence that I might be able to emulate the same process in order to figure it out as well. In the end, this focus is the ability that I developed in my 15 year gap since high school when I last tried programing. It is a decisive difference.

I could defeat the past me.